+++ src/main/kotlin/com/github/dinbtechit/vscodetheme/VSCodeThemeManager.kt
@@ -6,6 +6,12 @@ import com.intellij.ide.plugins.PluginManagerCore
 import com.intellij.ide.ui.LafManager
 import com.intellij.openapi.extensions.PluginId
 
+enum class VSCodeTheme(val theme: String) {
+    UNKNOWN("UNKNOWN"),
+    DARK("VSCode Dark"),
+    DARK_MODERN("VSCode Dark Modern");
+}
+
 class VSCodeThemeManager {
     companion object {
         const val pluginId = "com.github.dinbtechit.vscodetheme"
@@ -21,9 +27,9 @@ class VSCodeThemeManager {
     private fun getPlugin(): IdeaPluginDescriptor? = PluginManagerCore.getPlugin(PluginId.getId(pluginId))
     fun isVSCodeThemeReady(): Boolean {
         try {
-
             if (getPlugin()?.isEnabled != null) {
-                val vscodeTheme = LafManager.getInstance().installedLookAndFeels.first { it.name == "VSCode Dark" }
+                val vscodeTheme =
+                    LafManager.getInstance().installedLookAndFeels.first { it.name == VSCodeTheme.DARK.theme }
                 return vscodeTheme != null
             }
             return false
@@ -32,13 +38,15 @@ class VSCodeThemeManager {
         }
     }
 
-    fun switchToVSCodeTheme(always: Boolean = false) {
+    fun switchToVSCodeTheme(always: Boolean = false, selectedVSCodeTheme: VSCodeTheme = VSCodeTheme.DARK) {
         if (isVSCodeThemeReady()) {
-            val vscodeTheme = LafManager.getInstance().installedLookAndFeels.first { it.name == "VSCode Dark" }
+            val vscodeTheme =
+                LafManager.getInstance().installedLookAndFeels.first { it.name == selectedVSCodeTheme.theme }
             LafManager.getInstance().currentLookAndFeel = vscodeTheme
             if (always) {
                 val settings = VSCodeThemeSettingsStore.instance
                 settings.alwaysApply = true
+                settings.themeName = selectedVSCodeTheme
             }
         }
     }