+++ src/main/kotlin/com/github/dinbtechit/vscodetheme/startup/VSCodeStartupNotifyActivity.kt
@@ -1,8 +1,8 @@
 package com.github.dinbtechit.vscodetheme.startup
 
+import com.github.dinbtechit.vscodetheme.VSCodeTheme
 import com.github.dinbtechit.vscodetheme.VSCodeThemeManager
 import com.github.dinbtechit.vscodetheme.actions.AlwaysApplyThemeAction
-import com.github.dinbtechit.vscodetheme.actions.ApplyThemeAction
 import com.github.dinbtechit.vscodetheme.actions.DonateAction
 import com.github.dinbtechit.vscodetheme.actions.StarGithubRepoAction
 import com.github.dinbtechit.vscodetheme.icons.VSCodeIcons
@@ -13,11 +13,20 @@ import com.intellij.ide.ui.LafManager
 import com.intellij.notification.Notification
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.DefaultActionGroup
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
 
+
+
 class VSCodeStartupNotifyActivity : StartupActivity {
+
+    enum class DisplayActionType {
+        DONATION_ONLY,
+        SHOW_ALL_THEMES_FOR_DEFAULT,
+        SHOW_NEW_DARK_MODERN_THEME
+    }
 
     private val updateContent: String by lazy {
         //language=HTML
@@ -29,7 +38,16 @@ class VSCodeStartupNotifyActivity : StartupActivity {
     private val switchThemeQuestion: String by lazy {
         //language=HTML
         """
-          Set <b>VSCode Theme</b> as default theme.
+          Set one of the <b>VSCode Theme(s)</b> as a default theme.
+          <br/>
+          $updateContent
+        """.trimIndent()
+    }
+
+    private val tryNewDarkModernThemeQuestion: String by lazy {
+        //language=HTML
+        """
+          Would you like to set the <b>VSCode Dark Modern</b> as a default theme?
           <br/>
           $updateContent
         """.trimIndent()
@@ -38,6 +56,7 @@ class VSCodeStartupNotifyActivity : StartupActivity {
     companion object {
         const val pluginId = "com.github.dinbtechit.vscodetheme"
         lateinit var notification: Notification
+        var displayActionType: DisplayActionType = DisplayActionType.DONATION_ONLY
     }
 
     override fun runActivity(project: Project) {
@@ -46,11 +65,15 @@ class VSCodeStartupNotifyActivity : StartupActivity {
         if (isReady && getPlugin()?.version != VSCodeThemeSettingsStore.instance.version) {
             settings.version = getPlugin()!!.version
             if (settings.alwaysApply) {
-                VSCodeThemeManager.getInstance().switchToVSCodeTheme()
+                VSCodeThemeManager.getInstance().switchToVSCodeTheme(selectedVSCodeTheme = settings.themeName)
+                showNotificationPopup(project)
             } else if (settings.showNotificationOnUpdate) {
-                showUpdate(project)
+                showNotificationPopup(project)
             }
         }
+        // Uncomment for Testing popup
+        // showNotificationPopup(project)
+
     }
 
     private fun updateMsg(): String {
@@ -62,15 +85,28 @@ class VSCodeStartupNotifyActivity : StartupActivity {
         }
     }
 
-    private fun isVSCodeThemeSelected() = LafManager.getInstance().currentLookAndFeel.name == "VSCode Dark"
+    private fun isVSCodeThemeSelected() = LafManager.getInstance().currentLookAndFeel.name == VSCodeTheme.DARK.theme
+    private fun isVSCodeDarkModernThemeSelected() = LafManager.getInstance().currentLookAndFeel.name == VSCodeTheme.DARK_MODERN.theme
 
-    private fun showUpdate(project: Project) {
+    private fun showNotificationPopup(project: Project) {
         notification = createNotification(
             updateMsg(),
-            if (!isVSCodeThemeSelected()) switchThemeQuestion else updateContent,
+            notificationContent(),
             NotificationType.INFORMATION
         )
         showFullNotification(project, notification)
+    }
+
+    private fun notificationContent(): String {
+        if (!isVSCodeThemeSelected() && !isVSCodeDarkModernThemeSelected()) {
+            displayActionType = DisplayActionType.SHOW_ALL_THEMES_FOR_DEFAULT
+            return switchThemeQuestion
+        } else if (isVSCodeThemeSelected()) {
+            displayActionType = DisplayActionType.SHOW_NEW_DARK_MODERN_THEME
+            return tryNewDarkModernThemeQuestion
+        }
+        displayActionType = DisplayActionType.DONATION_ONLY
+        return updateContent
     }
 
     private fun getPlugin(): IdeaPluginDescriptor? = PluginManagerCore.getPlugin(PluginId.getId(pluginId))
@@ -83,9 +119,13 @@ class VSCodeStartupNotifyActivity : StartupActivity {
             .createNotification(content, type)
             .setTitle(title)
             .setIcon(VSCodeIcons.Logo).apply {
-                if (!isVSCodeThemeSelected()) {
-                    addAction(AlwaysApplyThemeAction())
-                    addAction(ApplyThemeAction())
+                if (displayActionType == DisplayActionType.SHOW_ALL_THEMES_FOR_DEFAULT) {
+                    addAction(DefaultActionGroup("Show All", false).apply {
+                        add(AlwaysApplyThemeAction(text = VSCodeTheme.DARK_MODERN.theme , vscodeTheme = VSCodeTheme.DARK_MODERN))
+                        add(AlwaysApplyThemeAction(text = VSCodeTheme.DARK.theme, vscodeTheme = VSCodeTheme.DARK))
+                    })
+                } else if (displayActionType == DisplayActionType.SHOW_NEW_DARK_MODERN_THEME) {
+                    addAction(AlwaysApplyThemeAction(text = "Switch Now" , vscodeTheme = VSCodeTheme.DARK_MODERN))
                 }
             }
             .addAction(DonateAction())