+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # vscode-theme Changelog
 
 ## Unreleased
+- #59 - VSCode Dark Modern Theme Initial
+- [Chores] - Updating links
 
 ## 1.7.9 - 2023-05-07
 - #55 - Syntax highlighting for shell scripts (For local variables explicitly)